requirements.txt.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
wikipedia-api>=0.6.0
colorama>=0.4.6
